Fix typo

+++ run_onchange_install-packages.sh
@@ -3,12 +3,14 @@
 sudo mkdir -p /etc/apt/keyrings
 
 # Add repository for eza
-wget -qO- https://raw.githubusercontent.com/eza-community/eza/main/deb.asc | sudo gpg --dearmor -o /etc/apt/keyrings/gierens.gpg
-echo "deb [signed-by=/etc/apt/keyrings/gierens.gpg] http://deb.gierens.de stable main" | sudo tee /etc/apt/sources.list.d/gierens.list
+if ! [ -f /etc/apt/keyrings/gierens.gpg ]; then
+  wget -qO- https://raw.githubusercontent.com/eza-community/eza/main/deb.asc | sudo gpg --dearmor -o /etc/apt/keyrings/gierens.gpg
+  echo "deb [signed-by=/etc/apt/keyrings/gierens.gpg] http://deb.gierens.de stable main" | sudo tee /etc/apt/sources.list.d/gierens.list
+fi
 
 sudo apt update
 
-sudo apt isntall -y \
+sudo apt install -y \
   curl \
   wget \
   fish \
